Refactor editor code into editor package
