+++ src/main/java/com/ezasm/gui/Theme.java
@@ -11,13 +11,11 @@ import java.awt.Font;
 /**
  * Represents a theme for components and text in the application.
  */
-public record Theme (Color background, Color foreground, Color currentLine, Color selection, Color comment,
-                     Color cyan, Color green, Color orange, Color pink, Color purple, Color red, Color yellow,
-                     boolean isLight) {
+public record Theme(Color background, Color foreground, Color currentLine, Color selection, Color comment, Color cyan,
+        Color green, Color orange, Color pink, Color purple, Color red, Color yellow, boolean isLight) {
 
     // based on https://github.com/dracula/dracula-theme
-    public static Theme Dracula = new Theme(
-            new Color(0x282a36), // background
+    public static Theme Dracula = new Theme(new Color(0x282a36), // background
             new Color(0xf8f8f2), // foreground
             new Color(0x44475a), // currentLine
             new Color(0x44475a), // selection
@@ -32,8 +30,7 @@ public record Theme (Color background, Color foreground, Color currentLine, Colo
             false); // is a light theme
 
     // based on https://github.com/endormi/vscode-2077-theme
-    public static Theme Purple = new Theme(
-            new Color(0x030d22), // background
+    public static Theme Purple = new Theme(new Color(0x030d22), // background
             new Color(0xfdfeff), // foreground
             new Color(0x310072), // currentLine
             new Color(0x35008b), // selection
@@ -48,8 +45,7 @@ public record Theme (Color background, Color foreground, Color currentLine, Colo
             false); // is a light theme
 
     // based on https://github.com/atom/one-light-syntax
-    public static Theme Light = new Theme(
-            new Color(0xebf8ff), // background
+    public static Theme Light = new Theme(new Color(0xebf8ff), // background
             new Color(0x161b1d), // foreground
             new Color(0xc1e4f6), // currentLine
             new Color(0x7ea2b4), // selection
