+++ src/main/java/com/ezasm/instructions/impl/TerminalInstructions.java
@@ -4,7 +4,7 @@ import java.io.*;
 import java.util.Scanner;
 import java.util.regex.Pattern;
 
-import com.ezasm.Conversion;
+import com.ezasm.util.Conversion;
 import com.ezasm.instructions.Instruction;
 import com.ezasm.instructions.targets.input.IAbstractInput;
 import com.ezasm.instructions.targets.output.IAbstractOutput;
