+++ src/main/java/com/ezasm/util/Arguments.java
@@ -1,4 +1,4 @@
-package com.ezasm;
+package com.ezasm.util;
 
 import com.ezasm.gui.Config;
 import com.ezasm.gui.Window;
@@ -6,8 +6,6 @@ import com.ezasm.simulation.ISimulator;
 import com.ezasm.simulation.Memory;
 import com.ezasm.simulation.Simulator;
 import org.apache.commons.cli.*;
-
-import javax.swing.*;
 
 /**
  * Methods to handle the program arguments and begin the program correspondingly.
